airlinesviewweb/airlinesviewweb/requirements.txt
flask
requests
pandas
python-dotenv
